% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{french_city}
\alias{french_city}
\title{French City Population}
\description{
Data frame containing latitude, longitude, population, and region code of
58 French cities.
}
\references{
\url{https://simplemaps.com/data/world-cities}
}
\keyword{data}
