% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{mapview}
\alias{mapview}
\title{Set Scales for Map View}
\usage{
mapview()
}
\description{
Add this as a layer to a ggplot object, along with \code{theme_void}, in
order to show only the map and none of the axes.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
