% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maptile.R
\name{mt_map_extent_data}
\alias{mt_map_extent_data}
\title{Compute and Download Map Tiles for Bounding Box}
\usage{
mt_map_extent_data(
  xmin,
  xmax,
  ymin,
  ymax,
  zoom = NULL,
  zoom_adj = 0,
  aspect_ratio = NULL,
  cache_dir = NULL,
  alpha = 1,
  url = NULL,
  force = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{xmin}{minimum value of the bounding box to cover: degrees of
longitude}

\item{xmax}{maximum value of the bounding box to cover: degrees of
longitude}

\item{ymin}{minimum value of the bounding box to cover: degrees of
latitude}

\item{ymax}{maximum value of the bounding box to cover: degrees of
latitude}

\item{zoom}{integer giving the zoom value, from 0 (entire world) to
19; set to NULL to auto-compute an appropriate zoom level}

\item{zoom_adj}{integer giving an adjustment to the default zoom level;
negative values zoom out from the default and positive
values zoom into the default. Useful to avoid
accidentally downloading a large number of tiles (by
setting a very large zoom value for a region) and for
adjusting facetted plots. Ignored when an explicit zoom
is supplied.}

\item{aspect_ratio}{positive numeric value giving the desired aspect ratio of
the tiles. When set to NULL, the default, tiles are
choosen to cover the data points without regard to the
aspect ratio.}

\item{cache_dir}{string giving the location of a local path in which to
download the tile files; if set to NULL, will first check
for the option "mt_cache_dir"; if also missing,
will determine a suitable temporary location that
persistsduring the R session}

\item{alpha}{optacity of the tiles; default is 1 (solid).}

\item{url}{url of the tile server, given as a string with three
values of "\%d", which will be filled in with the zoom,
x, and y values; if missing, will taken from the option
"mt_url"; otherwise will be set to:
"http://tile.stamen.com/toner/\%d/\%d/\%d.png".}

\item{force}{logical; should tiles be downloaded if they already
exist? Defaults to FALSE. Useful if the url is changed
or there was a network issue with the tiles.}

\item{quiet}{logical; should a message be shown indicating what tiles
are being downloaded and used? Set to FALSE by default.}
}
\value{
Returns a data frame with one row per tile, containing
                    a bounding box of the tile and the local location where
                    the file is located.
}
\description{
Given a range of x and y values (which correspond to longitude and latitude,
respecitively), the function determines a set of map tiles that cover the
region. Tiles are downloaded from a URL and stored in a local cache. A
data frame suitable for plotting with \code{geom_image_bbox} is returned.
}
