% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paris_metro}
\alias{paris_metro}
\title{Paris Metro Stations}
\description{
Data frame containing latitude, longitude, name, line, and color for each
metro stop in the city of Paris.
}
\references{
\url{https://www.data.gouv.fr/fr/datasets/}
}
\keyword{data}
